лоты: редактирование и обновление

+++ app/Http/Controllers/LotController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Company;
 use App\Models\Lot;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Validation\Rule;
 
 class LotController extends Controller
@@ -54,37 +55,68 @@ class LotController extends Controller
         return redirect(route('lot.index'))->with('success', 'Лот добавлен');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\Lot  $lot
-     * @return \Illuminate\Http\Response
-     */
-    public function show(Lot $lot)
-    {
-        //
-    }
-
 
     public function edit(Lot $lot)
     {
-       dd('edit: ' . $lot->id);
+        if (Gate::denies('company_owner', $lot->user_id)) {
+            abort(403, 'Access denied');
+        }
+
+        if ($lot->accepted_bid_id) {
+            return redirect(route('lot.index'))->with('error', 'Редактирование запрещено, т.к. есть принятая ставка');
+        }
+
+        $companies = Company::owner(auth()->id())->get();
+
+        return view('lot.edit', [
+               'lot' => $lot,
+               'companies' => $companies
+        ]);
     }
 
 
     public function update(Request $request, Lot $lot)
     {
-       dd('update: ' . $lot->id);
+        if (Gate::denies('company_owner', $lot->user_id)) {
+            abort(403, 'Access denied');
+        }
+
+        if ($lot->accepted_bid_id) {
+            return redirect(route('lot.index'))->with('error', 'Редактирование запрещено, т.к. есть принятая ставка');
+        }
+
+        $this->validate($request, [
+            'company_id' => 'required|integer|exists:companies,id',
+            'operation_type' => ['required', Rule::in(Lot::typesList())],
+            'nomenclature' => 'required|max:3000',
+            'NDS' => ['required', Rule::in([0,10,18])],
+            'sum' => 'required|numeric',
+            'fee' => 'required|numeric|between: 2,99',
+        ]);
+
+        $lot->update([
+            'company_id' => $request->company_id,
+            'operation_type' => $request->operation_type,
+            'nomenclature' => $request->nomenclature,
+            'NDS' => $request->NDS,
+            'sum' => $request->sum,
+            'fee' => $request->fee,
+        ]);
+
+        return redirect(route('lot.index'))->with('success', 'Лот изменен');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Lot  $lot
-     * @return \Illuminate\Http\Response
-     */
+
     public function destroy(Lot $lot)
     {
+        if (Gate::denies('company_owner', $lot->user_id)) {
+            abort(403, 'Access denied');
+        }
+
+        if ($lot->accepted_bid_id) {
+            return redirect(route('lot.index'))->with('error', 'Удаление запрещено, т.к. есть принятая ставка');
+        }
+
         return $lot->delete();
     }
 }
